[Backend] refactor: move chat handler logic - move chatManager from main file to main router file - move chat routes to its own router file - mount chat routes in main router file
